Moves MiddleEllipsisText to stripe UI core.
